add functional augmentations
