+++ yucca/image_processing/transforms/sampling.py
@@ -1,4 +1,5 @@
 from yucca.image_processing.transforms.YuccaTransform import YuccaTransform
+from yucca.functional.transforms import downsample_label
 import numpy as np
 from skimage.transform import resize
 
@@ -17,28 +18,7 @@ class DownsampleSegForDS(YuccaTransform):
         pass
 
     def __downsample__(self, label, factors):
-        orig_type = label.dtype
-        orig_shape = label.shape
-        downsampled_labels = []
-        for factor in factors:
-            target_shape = np.array(orig_shape).astype(int)
-            for i in range(2, len(orig_shape)):
-                target_shape[i] *= factor
-            if np.all(target_shape == orig_shape):
-                downsampled_labels.append(label)
-            else:
-                canvas = np.zeros(target_shape)
-                for b in range(label.shape[0]):
-                    for c in range(label[b].shape[0]):
-                        canvas[b, c] = resize(
-                            label[b, c].astype(float),
-                            target_shape[2:],
-                            0,
-                            mode="edge",
-                            clip=True,
-                            anti_aliasing=False,
-                        ).astype(orig_type)
-                downsampled_labels.append(canvas)
+        downsampled_labels = downsample_label(label, factors)
         return downsampled_labels
 
     def __call__(self, packed_data_dict=None, **unpacked_data_dict):
